+++ source/consumers/Prebinned1D.cpp
@@ -1,5 +1,5 @@
 #include <consumers/Prebinned1D.h>
-#include <consumers/dataspaces/dense1d.h>
+#include <consumers/dataspaces/Dense1D.h>
 
 #include <core/util/logger.h>
