+++ source/consumers/TOF1DCorrelate.cpp
@@ -1,5 +1,5 @@
 #include <consumers/TOF1DCorrelate.h>
-#include <consumers/dataspaces/dense1d.h>
+#include <consumers/dataspaces/Dense1D.h>
 
 #include <core/util/logger.h>
