+++ source/consumers/TimeDelta1D.cpp
@@ -1,5 +1,5 @@
 #include <consumers/TimeDelta1D.h>
-#include <consumers/dataspaces/dense1d.h>
+#include <consumers/dataspaces/Dense1D.h>
 
 #include <core/util/logger.h>
