+++ source/consumers/dataspaces/Dense1D.h
